+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 All notable changes to this project's current version will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
+to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 This new release includes :
 
 ### Added
 
+- Categorize ISO currencies (bond market unit, fund, precious metals, National or supranational and other) (#70).
+
 ### Changed
 
 ### Fixed
+
+- Add missing currency UYW (Unidad Previsional) (#69).
 
 ### Deprecated